fix: replace error snackbar to layout
